Text Analyzer which solves the following problem.
The application must take as input a text file, analyze the text and
output a report.  The report should contain a count of how many times
each word occurs in the file.  The output should be sorted, with a
primary sort of word length, and a secondary ASCII sort.

This solution uses the TreeMap java collection, 
Solution without using java collections coming soon.